Add isort options.
This adds the following isort options:
* --isort-profile
* --isort-wrap-length
* --isort-multi-line-output
* --isort-known-third-party
* --isort-known-first-party
* --isort-known-local-folder
* --isort-skip-gitignore
* --isort-use-parentheses
* --isort-length-sort

The newer options do not override isort own defaults to help ensure profile consistency. The existing gray defaults stay the same.

Added unittests for the args.py module.
Fixed a bug in args.py where raising an `ArgumentError` was failing at runtime, replaced it with a `ValueError`.